% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{mae}
\alias{mae}
\title{A function for calculating Mean Average Error}
\usage{
mae(actual, predicted)
}
\arguments{
\item{actual}{vector of validation data}

\item{predicted}{vector of predicted data}
}
\value{
numeric of MAE
}
\description{
This function will calculate MAE
}
\examples{
mae(actual,predicted)
}
