% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{mse}
\alias{mse}
\title{A function for calculating Mean Square Error}
\usage{
mse(actual, predicted)
}
\arguments{
\item{actual}{vector of validation data}

\item{predicted}{vector of predicted data}
}
\value{
numeric of MSE
}
\description{
This function will calculate MSE
}
\examples{
mse(actual,predicted)
}
