% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.function.R
\name{OK_krig}
\alias{OK_krig}
\title{An Ordinary Kriging function}
\usage{
OK_krig(data, grid, data.val, property)
}
\arguments{
\item{data}{SpatialPointsDataFrame of calibration data}

\item{grid}{grid for kriging}

\item{data.val}{SpatialPointsDataFrame of validation data}

\item{property}{char of property name in 'data' for kriging. Case sensitive}
}
\value{
list of error indicies for each subsample size
}
\description{
This function will perform orindary kriging
}
