% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.function.R
\name{rep.function}
\alias{rep.function}
\title{A function for repetition of Oridinary Kriging}
\usage{
\method{rep}{`function`}(css, depth_spdf, grid, replications, property)
}
\arguments{
\item{css}{vector of subsample lengths}

\item{depth_spdf}{SpatialPointsDataFrame of data}

\item{grid}{grid to krige across}

\item{replications}{numeric of required replications}

\item{property}{char of property name in 'depth_spdf' for kriging. Case sensitive}
}
\value{
list of error indicies for each subsample size
}
\description{
This function will iterate through a vector of subsample sizes
}
