% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.function.rk.R
\name{rep.function.rk}
\alias{rep.function.rk}
\title{A function for repetition of Regression Kriging}
\usage{
\method{rep}{function.rk}(css, depth_spdf, grid, rasters, replications, layers)
}
\arguments{
\item{css}{vector of subsample lengths}

\item{depth_spdf}{SpatialPointsDataFrame of data}

\item{grid}{grid to krige across}

\item{rasters}{RasterStack of environmental covariate rasters}

\item{replications}{numeric of required replications}

\item{layers}{chr string of layer names for naming in output list}
}
\value{
list of error indicies for each subsample size
}
\description{
This function will iterate through a vector of subsample sizes
}
\examples{
rep.function.rk(vec,spdf,grid,rasters)
}
