% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.function.rk.R
\name{RK_krig}
\alias{RK_krig}
\title{A Regression Kriging function}
\usage{
RK_krig(data, grid, data.val, rasters, proj4Str, l)
}
\arguments{
\item{data}{SpatialPointsDataFrame of calibration data}

\item{grid}{grid for kriging}

\item{data.val}{SpatialPointsDataFrame of validation data}

\item{rasters}{RasterStack of environmental covariate rasters}

\item{proj4Str}{chr string of the projection of data}

\item{l}{numeric column index of variable for kriging}
}
\value{
list of error indicies for each subsample size
}
\description{
This function will perform regression kriging
}
\examples{
rep.function.rk(vec,spdf,grid,rasters)
}
