% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{A function for calculating Root Mean Square Error}
\usage{
rmse(actual, predicted)
}
\arguments{
\item{actual}{vector of validation data}

\item{predicted}{vector of predicted data}
}
\value{
numeric of RMSE
}
\description{
This function will calculate RMSE
}
\examples{
rmse(actual,predicted)
}
