% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rsq}
\alias{rsq}
\title{A function for calculating Rsquared}
\usage{
rsq(actual, predicted)
}
\arguments{
\item{actual}{vector of validation data}

\item{predicted}{vector of predicted data}
}
\value{
numeric of RSQ
}
\description{
This function will calculate RSQ
}
\examples{
rsq(actual,predicted)
}
