% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rsq_adj}
\alias{rsq_adj}
\title{A function for calculating Adjusted Rsquared}
\usage{
rsq_adj(actual, predicted, p)
}
\arguments{
\item{actual}{vector of validation data}

\item{predicted}{vector of predicted data}

\item{p}{numeric of number of predictors used}
}
\value{
numeric of RSQ
}
\description{
This function will calculate RSQ_ADJ
}
