% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{sde}
\alias{sde}
\title{A function for calculating Standard Deviation of residuals}
\usage{
sde(actual, predicted)
}
\arguments{
\item{actual}{vector of validation data}

\item{predicted}{vector of predicted data}
}
\value{
numeric of SDE
}
\description{
This function will calculate SDE
}
