% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.function.R
\name{OK_krig}
\alias{OK_krig}
\title{An Ordinary Kriging function}
\usage{
OK_krig(data, grid, data.val, proj4Str, l)
}
\arguments{
\item{data}{SpatialPointsDataFrame of calibration data}

\item{grid}{grid for kriging}

\item{data.val}{SpatialPointsDataFrame of validation data}

\item{proj4Str}{chr string of the projection of data}

\item{l}{numeric column index of variable for kriging}
}
\value{
list of error indicies for each subsample size
}
\description{
This function will perform orindary kriging
}
\examples{
rep.function.rk(vec,spdf,grid)
}
